+++ sdks/python/apache_beam/io/gcp/bigquery_tools.py
@@ -43,9 +43,10 @@ from typing import Sequence
 from typing import TypeVar
 from typing import Union
 
-import apache_beam
 import fastavro
 import numpy as np
+
+import apache_beam
 from apache_beam import coders
 from apache_beam.internal.gcp import auth
 from apache_beam.internal.gcp.json_value import from_json_value
@@ -367,16 +368,8 @@ class BigQueryWrapper(object):
     self.quota_project_id = quota_project_id
     self.client = client or BigQueryWrapper._bigquery_client(
         PipelineOptions(), quota_project_id=quota_project_id)
-
-    # If the client is a mock (common in tests) or has the specific method
-    # we use, we use it as the gcp_bq_client to preserve backward
-    # compatibility for tests. Otherwise (e.g. it's a real apitools client),
-    # we create the correct google-cloud-bigquery client.
-    if client and hasattr(client, 'insert_rows_json'):
-      self.gcp_bq_client = client
-    else:
-      self.gcp_bq_client = BigQueryWrapper._gcp_bigquery_client(
-          quota_project_id=quota_project_id)
+    self.gcp_bq_client = client or BigQueryWrapper._gcp_bigquery_client(
+        quota_project_id=quota_project_id)
 
     self._unique_row_id = 0
     # For testing scenarios where we pass in a client we do not want a